Add Question_Response Resource
Adds new resource Question_Response, which serves as an intermediary in
a many-to-many relationship between users and questions.

Question_Responses simply consist of a user_id and a question_id, and
act on the front end as a way to determine if a user has answered a
particular question before.

Adds create and index routes for question_responses in routes.

Adds CURL script to test create action.

Adds serializer to send back the user and question ids.